+++ 3rd-Party/ctk/include/dcmtk/dcmrt/seq/drtcdrs.h
@@ -1,13 +1,13 @@
 /*
  *
  *  Copyright (C) 2008-2012, OFFIS e.V. and ICSMED AG, Oldenburg, Germany
- *  Copyright (C) 2013-2016, J. Riesmeier, Oldenburg, Germany
+ *  Copyright (C) 2013-2017, J. Riesmeier, Oldenburg, Germany
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  Header file for class DRTCalculatedDoseReferenceSequence
  *
- *  Generated automatically from DICOM PS 3.3-2016d
- *  File created on 2016-10-12 13:44:31
+ *  Generated automatically from DICOM PS 3.3-2017e
+ *  File created on 2017-12-05 09:30:54
  *
  */
 
@@ -55,6 +55,7 @@ class DCMTK_DCMRT_EXPORT DRTCalculatedDoseReferenceSequence
 
         /** assigment operator
          *  @param copy item object to be copied
+         *  @return reference to this object
          */
         Item &operator=(const Item &copy);
 
@@ -203,6 +204,7 @@ class DCMTK_DCMRT_EXPORT DRTCalculatedDoseReferenceSequence
 
     /** assigment operator
      *  @param copy sequence object to be copied
+     *  @return reference to this object
      */
     DRTCalculatedDoseReferenceSequence &operator=(const DRTCalculatedDoseReferenceSequence &copy);
 
@@ -225,7 +227,7 @@ class DCMTK_DCMRT_EXPORT DRTCalculatedDoseReferenceSequence
     /** get number of items in the sequence
      *  @return number of items
      */
-    unsigned long getNumberOfItems() const;
+    size_t getNumberOfItems() const;
 
     /** goto first item in the sequence
      *  @return status, EC_Normal if successful, an error code otherwise
@@ -241,7 +243,7 @@ class DCMTK_DCMRT_EXPORT DRTCalculatedDoseReferenceSequence
      *  @param  num  number of the item to be selected (0..num-1)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition gotoItem(const unsigned long num);
+    OFCondition gotoItem(const size_t num);
 
     /** get current item in the sequence
      *  @param  item  reference to item pointer (result variable)
@@ -264,31 +266,31 @@ class DCMTK_DCMRT_EXPORT DRTCalculatedDoseReferenceSequence
      *  @param  item  reference to item pointer (result variable)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition getItem(const unsigned long num, Item *&item);
+    OFCondition getItem(const size_t num, Item *&item);
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return reference to specified item if successful, empty default item otherwise
      */
-    Item &getItem(const unsigned long num);
+    Item &getItem(const size_t num);
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return const reference to specified item if successful, empty default item otherwise
      */
-    const Item &getItem(const unsigned long num) const;
+    const Item &getItem(const size_t num) const;
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return reference to specified item if successful, empty default item otherwise
      */
-    Item &operator[](const unsigned long num);
+    Item &operator[](const size_t num);
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return const reference to specified item if successful, empty default item otherwise
      */
-    const Item &operator[](const unsigned long num) const;
+    const Item &operator[](const size_t num) const;
 
     /** add new item to the end of this sequence
      *  @param  item  reference to new item pointer (result variable)
@@ -301,13 +303,13 @@ class DCMTK_DCMRT_EXPORT DRTCalculatedDoseReferenceSequence
      *  @param  item  reference to new item pointer (result variable)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition insertItem(const unsigned long pos, Item *&item);
+    OFCondition insertItem(const size_t pos, Item *&item);
 
     /** remove particular item from the sequence
      *  @param  pos  position of the item to be removed (0..num-1)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition removeItem(const unsigned long pos);
+    OFCondition removeItem(const size_t pos);
 
   // --- input/output methods ---
 
@@ -342,7 +344,7 @@ class DCMTK_DCMRT_EXPORT DRTCalculatedDoseReferenceSequence
      *  @param  iterator  list iterator storing the position of the item
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition gotoItem(const unsigned long num,
+    OFCondition gotoItem(const size_t num,
                          OFListIterator(Item *) &iterator);
 
     /** goto particular item in the sequence
@@ -350,7 +352,7 @@ class DCMTK_DCMRT_EXPORT DRTCalculatedDoseReferenceSequence
      *  @param  iterator  list iterator storing the position of the item
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition gotoItem(const unsigned long num,
+    OFCondition gotoItem(const size_t num,
                          OFListConstIterator(Item *) &iterator) const;
 
   private:
