+++ 3rd-Party/ctk/include/dcmtk/dcmrt/seq/drtipiqs.h
@@ -1,13 +1,13 @@
 /*
  *
  *  Copyright (C) 2008-2012, OFFIS e.V. and ICSMED AG, Oldenburg, Germany
- *  Copyright (C) 2013-2016, J. Riesmeier, Oldenburg, Germany
+ *  Copyright (C) 2013-2017, J. Riesmeier, Oldenburg, Germany
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  Header file for class DRTIssuerOfPatientIDQualifiersSequence
  *
- *  Generated automatically from DICOM PS 3.3-2016d
- *  File created on 2016-10-12 13:44:31
+ *  Generated automatically from DICOM PS 3.3-2017e
+ *  File created on 2017-12-05 09:30:54
  *
  */
 
@@ -58,6 +58,7 @@ class DCMTK_DCMRT_EXPORT DRTIssuerOfPatientIDQualifiersSequence
 
         /** assigment operator
          *  @param copy item object to be copied
+         *  @return reference to this object
          */
         Item &operator=(const Item &copy);
 
@@ -213,6 +214,7 @@ class DCMTK_DCMRT_EXPORT DRTIssuerOfPatientIDQualifiersSequence
 
     /** assigment operator
      *  @param copy sequence object to be copied
+     *  @return reference to this object
      */
     DRTIssuerOfPatientIDQualifiersSequence &operator=(const DRTIssuerOfPatientIDQualifiersSequence &copy);
 
@@ -235,7 +237,7 @@ class DCMTK_DCMRT_EXPORT DRTIssuerOfPatientIDQualifiersSequence
     /** get number of items in the sequence
      *  @return number of items
      */
-    unsigned long getNumberOfItems() const;
+    size_t getNumberOfItems() const;
 
     /** goto first item in the sequence
      *  @return status, EC_Normal if successful, an error code otherwise
@@ -251,7 +253,7 @@ class DCMTK_DCMRT_EXPORT DRTIssuerOfPatientIDQualifiersSequence
      *  @param  num  number of the item to be selected (0..num-1)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition gotoItem(const unsigned long num);
+    OFCondition gotoItem(const size_t num);
 
     /** get current item in the sequence
      *  @param  item  reference to item pointer (result variable)
@@ -274,31 +276,31 @@ class DCMTK_DCMRT_EXPORT DRTIssuerOfPatientIDQualifiersSequence
      *  @param  item  reference to item pointer (result variable)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition getItem(const unsigned long num, Item *&item);
+    OFCondition getItem(const size_t num, Item *&item);
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return reference to specified item if successful, empty default item otherwise
      */
-    Item &getItem(const unsigned long num);
+    Item &getItem(const size_t num);
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return const reference to specified item if successful, empty default item otherwise
      */
-    const Item &getItem(const unsigned long num) const;
+    const Item &getItem(const size_t num) const;
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return reference to specified item if successful, empty default item otherwise
      */
-    Item &operator[](const unsigned long num);
+    Item &operator[](const size_t num);
 
     /** get particular item in the sequence
      *  @param  num  number of the item to be retrieved (0..num-1)
      *  @return const reference to specified item if successful, empty default item otherwise
      */
-    const Item &operator[](const unsigned long num) const;
+    const Item &operator[](const size_t num) const;
 
     /** add new item to the end of this sequence
      *  @param  item  reference to new item pointer (result variable)
@@ -311,13 +313,13 @@ class DCMTK_DCMRT_EXPORT DRTIssuerOfPatientIDQualifiersSequence
      *  @param  item  reference to new item pointer (result variable)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition insertItem(const unsigned long pos, Item *&item);
+    OFCondition insertItem(const size_t pos, Item *&item);
 
     /** remove particular item from the sequence
      *  @param  pos  position of the item to be removed (0..num-1)
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition removeItem(const unsigned long pos);
+    OFCondition removeItem(const size_t pos);
 
   // --- input/output methods ---
 
@@ -352,7 +354,7 @@ class DCMTK_DCMRT_EXPORT DRTIssuerOfPatientIDQualifiersSequence
      *  @param  iterator  list iterator storing the position of the item
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition gotoItem(const unsigned long num,
+    OFCondition gotoItem(const size_t num,
                          OFListIterator(Item *) &iterator);
 
     /** goto particular item in the sequence
@@ -360,7 +362,7 @@ class DCMTK_DCMRT_EXPORT DRTIssuerOfPatientIDQualifiersSequence
      *  @param  iterator  list iterator storing the position of the item
      *  @return status, EC_Normal if successful, an error code otherwise
      */
-    OFCondition gotoItem(const unsigned long num,
+    OFCondition gotoItem(const size_t num,
                          OFListConstIterator(Item *) &iterator) const;
 
   private:
